Set env var on devcontainer needed for unit tests
There is currently an `init` function in the `util` package which
requires the `GAUGE_PROJECT_ROOT` environment variable to be set.  If
this environment variable is not set then the init function errors when
running unit tests, causing the tests to fail.  This commit fixes the
issue when running locally in a devcontainer or GitHub Codespace, by
setting the environment variable in the container's configuration.