Make binaries available as GH Action artifacts
The standard way for users of the plugin to use the binaries is to run
`gauge install jira`.  Plus the binaries are also available via [GitHub
Releases][1].  This commit adds a third way to access the binaries, by
making them available via a GitHub Action's artifacts.  This can be
useful for installing an experimental version on a branch that has not
yet been released, for instance.

[1]: https://docs.github.com/github/administering-a-repository/about-releases