Fix issue where builds were running twice
The builds were running on every push and every pull request
change, meaning that when opening or editing a pull request they were
running twice, at the same time.  This could lead to unexpected
results.

Solution is to [run them on all pushes but only on pull requests which
come from a fork][1].

[1]: https://github.community/t/how-to-trigger-an-action-on-push-or-pull-request-but-not-both/16662/10?