Fix release process - now gauge-bot independent
This repo was generated from the [getgauge/spectacle][1] repo, which
uses the Gauge org's own [gauge-bot][2] to create releases.  Our repo
is independent of the Gauge org, so this commit enables [releases][3] to
be created independently of the bot.

To create a release, run the `release.sh` script.

NB if you are using a [GitHub Codespace][4] as your IDE, the GitHub
token that is automatically set by GitHub in the Codespace doesn't have
enough permissions to run the script successfully.  So you will need to
use a [GitHub personal access token][5] that has `public_repo` scope,
e.g.

`GITHUB_TOKEN=blah ./release.sh`

The changes in this commit are based on the [Spectacle repo's previous
way of releasing][6] before it started to use the Gauge bot.

[1]: https://github.com/getgauge/spectacle
[2]: https://github.com/getgauge/gauge-bot
[3]: https://docs.github.com/github/administering-a-repository/about-releases
[4]: https://github.com/features/codespaces
[5]: https://docs.github.com/en/github/authenticating-to-github/creating-a-personal-access-token
[6]: https://github.com/getgauge/spectacle/blob/bea306ffe848ad6cee89a53643a3ba3ec1b86c05/.github/workflows/github_release.yml