Make replacement of specs in Jira less brittle
Prior to this commit the replacement of specs in Jira was coupled to
the subheader message as well as to the header.  This meant that
if we changed the subheader message in the future then the specs already
published in Jira would not be replaced as they should be.  This commit
fixes this issue by using just the header to match against, meaning that
now we can freely change the subheader message with no side-effects.