Add spec replacement contract test
It is important that the contract to replace Gauge specifications in the
Jira issue descriptions does not change.  It would be a breaking change
if it did, as it would mean that existing users with Gauge specs already
published to Jira would not have these existing specifications replaced
when running the plugin.

This commit therefore adds a test to verify that the contract for
replacing Gauge specifications in the Jira issue descriptions remains
unchanged.