Test entire console output to strengthen tests
Checking the entire console output rather than just part of it is gives
us more regression coverage.

The console output sometimes differs on Jira Cloud from Jira Server,
because for instance the error messages logged on the console differ. So
we check the entire console output only for those scenarios which do not
differ on Jira Cloud and Server.