Handle Jira field exceeding max length (#15)
* Handle Jira field exceeding max length

This commit adds information to the console output when spec publishing
fails for an individual Jira issue because the specs cause the maximum
length of the Jira description field to be exceeded.

The default maximum field length in Jira is 32767 characters.  We check
if the length exceeds this and display a message on the console if so,
advising how to change the maximum length.

NB The Jira API does not allow us to query the current maximum length
currently set in the Jira instance (it is a configurable value, so
it may be different to the default).  This means that our message
displayed on the console will be somewhat misleading in the case where
the maximum field length has been changed from the default.  This
scenario should be rare though, so no need to handle it more gracefully
at this moment in time.

* Bump plugin version to 0.1.3

Also formatted the plugin.json file