Prevent publishing when issue is in invalid state
Jira issues should only ever contain one Gauge examples section,
currently (NB we may change this in the future if we cater for more than
one source repo separately publishing their Gauge specs to the same
Jira issue).

This commit prevents publishing of specs to a Jira issue which is in an
invalid state because it has more than one Gauge examples section in the
Jira issue description.

A Jira issue could get into this invalid state either because of a
manual edit, or because of a bug (hypothetically) in the Gauge Jira
plugin itself which inadvertently led to a duplicate examples section
instead of replacing the existing one.