Fix duplicate Gauge examples section bug
Sometimes when a Jira issue is edited manually it will lead to headings
(e.g. in the description field) either having a space or not having a
space between the heading and the heading text.
The bug was caused by us always expecting the Gauge specs section in the
Jira to have an h2 heading with no space between the h2 and the heading
text.  This commit fixes that bug by handling both scenarios - we now
allow for there to be no space, or any number of spaces.