Make logic to replace Gauge section less brittle
This commit caters for a situation where the whitespace in the Gauge
specifications section of the Jira issue description changes (e.g. if
a manual edit to the issue description inadvertently changes it).
Prior to this commit that would cause the Gauge specs section to be
duplicated (rather than being replaced as it should), on the next
running of the Gauge Jira plugin.