Fix duplicate Gauge examples section bug (#20)
* Bump plugin version to 0.1.7

* Fix duplicate Gauge examples section bug

Sometimes when a Jira issue is edited manually it will lead to headings
(e.g. in the description field) either having a space or not having a
space between the heading and the heading text.
The bug was caused by us always expecting the Gauge specs section in the
Jira to have an h2 heading with no space between the h2 and the heading
text.  This commit fixes that bug by handling both scenarios - we now
allow for there to be no space, or any number of spaces.

* Make logic to replace Gauge section less brittle

This commit caters for a situation where the whitespace in the Gauge
specifications section of the Jira issue description changes (e.g. if
a manual edit to the issue description inadvertently changes it).
Prior to this commit that would cause the Gauge specs section to be
duplicated (rather than being replaced as it should), on the next
running of the Gauge Jira plugin.

* Fix bug where footer was sometimes indented

Sometimes the footer of the Gauge specs section in Jira was being
erroneously indented, e.g. if the specs finished with a table.
This commit fixes the bug by adding a newline at the beginning of the
footer.

* Extract method to new regex utility package