Fail fast if multiple spec directories passed in
Gauge documentation plugins allow multiple specs directories to be
passed in when invoking the plugin (from the command line).  We now fail
fast if more than one directory is passed in, as this Jira plugin needs
there to be just one specs directory (so it can be mapped to the
specs Git url).